pyproject.toml
[tool.poetry]
name = "ymldb"
version = "0.1.0"
description = "A file based, NoSQL database with atomic change rollbacks."
authors = ["Dan Trickey <dan@trickey.io>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/pyinv/ymldb"

[tool.poetry.dependencies]
python = "^3.6"
pydantic = "^0.32.1"
attrs = "^19.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.7"
flake8-commas = "^2.0"
flake8-comprehensions = "^2.2"
flake8-debugger = "^3.1"
flake8-docstrings = "^1.3"
flake8-mutable = "^1.2"
flake8-todo = "^0.7.0"
mypy = "^0.720.0"
pytest = "^5.1"
pytest-cov = "^2.7"
rope = "^0.14.0"
isort = {version = "^4.3", extras = ["pyproject"]}
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
